updated wishlist module
